+++ src/stream/mod.rs
@@ -15,6 +15,7 @@ use serde::Serialize;
 
 pub mod broadcast;
 pub mod event_log;
+pub mod redact;
 pub mod sse;
 #[cfg(feature = "webhook")]
 pub mod sweep_webhook;
